+++ library/src/lapack/roclapack_getri_optimized.cpp
@@ -30,7 +30,9 @@ ROCSOLVER_KERNEL void __launch_bounds__(WAVESIZE) getri_kernel_small(U AA,
 
     // batch instance
     T* A = load_ptr_batch<T>(AA, b, shiftA, strideA);
-    rocblas_int* ipiv = load_ptr_batch<rocblas_int>(ipivA, b, shiftP, strideP);
+    rocblas_int* ipiv;
+    if(ipivA != nullptr)
+        ipiv = load_ptr_batch<rocblas_int>(ipivA, b, shiftP, strideP);
 
     // shared memory (for communication between threads in group)
     __shared__ T common[DIM];
@@ -116,15 +118,18 @@ ROCSOLVER_KERNEL void __launch_bounds__(WAVESIZE) getri_kernel_small(U AA,
     }
 
     // apply pivots (getri_pivot)
-#pragma unroll
-    for(rocblas_int j = DIM - 2; j >= 0; j--)
+    if(ipivA != nullptr)
     {
-        jp = ipiv[j] - 1;
-        if(jp != j)
+#pragma unroll
+        for(rocblas_int j = DIM - 2; j >= 0; j--)
         {
-            temp = rA[j];
-            rA[j] = rA[jp];
-            rA[jp] = temp;
+            jp = ipiv[j] - 1;
+            if(jp != j)
+            {
+                temp = rA[j];
+                rA[j] = rA[jp];
+                rA[jp] = temp;
+            }
         }
     }