+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Full documentation for rocSOLVER is available at [rocsolver.readthedocs.io](http
 
 ## [(Unreleased) rocSOLVER]
 ### Added
+- General matrix inversion without pivoting:
+    - GETRI\_NPVT (with batched and strided\_batched versions)
+    - GETRI\_NPVT\_OUTOFPLACE (with batched and strided\_batched versions)
 - Added `rocsolver_get_version_string_size` to help with version string queries
 
 ### Optimized